pyproject.toml
[project]
name = "googleadk-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "elasticsearch==8.15.1",
    "fastmcp>=2.12.4",
    "google-adk>=1.15.1",
    "numpy>=2.3.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
]
